requirements.txt
sqlite-utils~=1.11
GitPython~=3.0.2
datasette
datasette-vega
datasette-publish-vercel
